pyproject.toml
[tool.poetry]
name = "scrfd"
version = "0.1.0"
description = "SCRFD inference"
authors = ["cospectrum"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.1"
onnxruntime = "^1.16.1"
pillow = "^10.1.0"
pydantic = "^2.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
mypy = "^1.6.1"
ruff = "^0.1.4"
pyright = "^1.1.334"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
